Replaced Student object with user. Refactored most functions to handle that change. Added tons of print statements to trackdown what is and isn't working.
Currently, it is navigating the search space. Unfortunately, it hasnt' hit a goal state on any attempts. I think we are having a combination of problems at the moment: Our query is likely giving us duplicate courses in the same term. I'm not convinced that my addCourse method in the Plan object is working correctly. It seems to add a lot of empty lists. The filters appear to be working correctly.

Lots more testing and fiddling required to get this working.